requirements.txt
absl-py==2.3.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
alembic==1.16.5
annotated-types==0.7.0
astunparse==1.6.3
attrs==25.3.0
binance-connector==3.12.0
binance-futures-connector==4.1.0
certifi==2025.8.3
charset-normalizer==3.4.3
cloudpickle==3.1.1
colorama==0.4.6
colorlog==6.9.0
contourpy==1.3.3
cycler==0.12.1
datasets==4.0.0
dill==0.3.8
Farama-Notifications==0.0.4
filelock==3.19.1
flatbuffers==25.2.10
fonttools==4.59.2
frozenlist==1.7.0
fsspec==2025.3.0
gast==0.6.0
google-pasta==0.2.0
greenlet==3.2.4
grpcio==1.74.0
gymnasium==0.29.1
h5py==3.14.0
huggingface-hub==0.34.4
idna==3.10
Jinja2==3.1.6
joblib==1.5.2
keras==3.11.3
kiwisolver==1.4.9
libclang==18.1.1
Mako==1.3.10
Markdown==3.8.2
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib==3.10.6
mdurl==0.1.2
ml_dtypes==0.5.3
mpmath==1.3.0
multidict==6.6.4
multiprocess==0.70.16
namex==0.1.0
networkx==3.5
numpy==2.3.2
opt_einsum==3.4.0
optree==0.17.0
optuna==4.5.0
packaging==25.0
pandas==2.3.2
pillow==11.3.0
propcache==0.3.2
protobuf==6.32.0
pyarrow==21.0.0
pycryptodome==3.23.0
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
requests==2.32.5
rich==14.1.0
scikit-learn==1.7.1
scipy==1.16.1
seaborn==0.13.2
setuptools==80.9.0
six==1.17.0
SQLAlchemy==2.0.43
sympy==1.14.0
ta==0.11.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow_cpu==2.20.0
termcolor==3.1.0
threadpoolctl==3.6.0
torch==2.8.0+cu126
torchvision==0.23.0+cu126
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
websocket-client==1.8.0
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.3
xxhash==3.5.0
yarl==1.20.1
